[ADD Update Product

+++ src/main/java/com/example/spring/model/ProductModel.java
@@ -53,7 +53,7 @@ public class ProductModel {
     public void setDescription(String description) {this.description = description;}
 
     public double getPrice() {return price;}
-    public void setPrice(int price) {this.price = price;}
+    public void setPrice(double price) {this.price = price;}
 
     public int getQuantity() {return quantity;}
     public void setQuantity(int quantity) {this.quantity = quantity;}
